added so that load_corrections uses different pedestals for different runs, fixed bug with passing threshold

+++ scripts/process.py
@@ -27,7 +27,7 @@ from integrators import *
 parser = argparse.ArgumentParser(description='Analyze a run of p17743 at SwissFEL.')
 parser.add_argument('-r', '--run', type=str, required=True, help='run number to process')
 parser.add_argument('-s','--shots',type=int, default=0, help='number of shots to process (default: 0 = all)')
-parser.add_argument('-t','--threshold',type=int, default=0, help='hit threshold for radial profile (default: 0 = no hits)')
+parser.add_argument('-t','--threshold',type=float, default=0, help='hit threshold for radial profile (default: 0 = no hits)')
 parser.add_argument('-p','--path',type=str,default='/sf/bernina/data/p17743/res/scan_info/',help='path to data')
 #'/sf/bernina/data/p17743/scratch/hdf5'
